+++ src/Processors/QueryPlan/FractionalOffsetStep.cpp
@@ -29,7 +29,7 @@ static ITransformingStep::Traits getTraits()
     };
 }
 
-FractionalOffsetStep::FractionalOffsetStep(const SharedHeader & input_header_, Float32 fractional_offset_)
+FractionalOffsetStep::FractionalOffsetStep(const SharedHeader & input_header_, Float64 fractional_offset_)
     : ITransformingStep(input_header_, input_header_, getTraits())
     , fractional_offset(fractional_offset_)
 {
@@ -59,7 +59,7 @@ void FractionalOffsetStep::serialize(Serialization & ctx) const
 
 std::unique_ptr<IQueryPlanStep> FractionalOffsetStep::deserialize(Deserialization & ctx)
 {
-    Float32 offset;
+    Float64 offset;
     readFloatBinary(offset, ctx.in);
 
     return std::make_unique<FractionalOffsetStep>(ctx.input_headers.front(), offset);
